// src/api/generated/model/slackUserResponseDto.ts
/**
 * Generated by orval v8.4.1 🍺
 * Do not edit manually.
 * Dev Blackbox API
 * OpenAPI spec version: 0.1.0
 */

export interface SlackUserResponseDto {
  id: number;
  member_id: string;
  display_name: string;
  real_name: string;
  email: string | null;
  user_id: number | null;
  created_at: string;
  updated_at: string;
}